+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.7.0] - 2026-03-01
+
+### Changed
+- `install.sh` now downloads from GitHub Releases
+- Self-update now queries the GitHub Releases API for new versions
+- Releases are published to both Gitea and GitHub
+
 ## [0.6.0] - 2026-03-01
 
 ### Added